refactor: moved position file to 'src' root
