+++ src/frontend/lexer.ts
@@ -11,7 +11,7 @@
  */
 
 import { basename } from "@std/path/basename"
-import { Position } from "@frontend/position.ts"
+import { Position } from "@/position.ts"
 import { KEYWORD_MAP, TOKEN_MAP, RTOKEN_MAP, Token, Tokens } from "@frontend/tokens/tokens.ts"
 import { Collectable, Reader } from "@frontend/typings.ts"
